Fix: 게시글 api - null값 발생시 error handling

+++ back-end/backend/src/main/java/com/groot/backend/service/ArticleServiceImpl.java
@@ -38,7 +38,7 @@ public class ArticleServiceImpl implements ArticleService{
     private final RedisTemplate<String, String> redisTemplate;
     private final TagCountRepository tagCountRepository;
 
-
+    // 지역명 리스트 조회
     @Override
     public List<String> readRegion() {
         List<RegionEntity> regions = regionRepository.findAll();
@@ -49,6 +49,7 @@ public class ArticleServiceImpl implements ArticleService{
         return result;
     }
 
+    // 게시글 존재 여부 확인
     @Override
     public boolean existedArticleId(Long articleId) {
         return articleRepository.existsById(articleId);
@@ -92,7 +93,6 @@ public class ArticleServiceImpl implements ArticleService{
         }else{
             // 조회되는 값은 있는데, 리셋 후 입력된 태그가 없으면 전날 데이터를 보여줌
             if(Double.compare(result.get(0).getCount(), 0.0) == 0){
-                System.out.println("전날데이터");
                 List<TagCountEntity> list = tagCountRepository.findAll(Sort.by(Sort.Direction.DESC, "count"));
 
                 // tagCount 데이터가 존재하지 않으면 태그 데이터만 보여줌
@@ -234,28 +234,32 @@ public class ArticleServiceImpl implements ArticleService{
         ArticleEntity articleEntity = articleRepository.findById(articleId).orElseThrow();
         // UserEntity 조회
         UserEntity userEntity = userRepository.findById(articleEntity.getUserPK()).orElseThrow();
+
         // tags id 조회
         List<ArticleTagEntity> articleTagEntityList = articleTagRepository.findByArticleId(articleId);
-        // tag id로 tagEntity 조회
         List<String> tags = new ArrayList<>();
-        for(ArticleTagEntity articleTagEntity : articleTagEntityList){
-            tags.add(tagRepository.findById(articleTagEntity.getTagId()).orElseThrow().getName());
+        if(articleTagEntityList.size() != 0){
+            // tag id로 tagEntity 조회
+            for(ArticleTagEntity articleTagEntity : articleTagEntityList){
+                tags.add(tagRepository.findById(articleTagEntity.getTagId()).orElseThrow().getName());
+            }
         }
+
         // commentEntity 조회
         List<CommentEntity> commentEntityList = commentRepository.findByArticleId(articleId);
 
         List<CommentResponseDTO> comments = new ArrayList<>();
         int commentCnt = 0;
         // commentDTO build
-        if(commentEntityList != null){
+        if(commentEntityList.size() != 0){
             for(CommentEntity commentEntity : commentEntityList){
                 UserEntity commentUserEntity = userRepository.findById(commentEntity.getUserPK()).orElseThrow();
                 CommentResponseDTO commentDTO = CommentResponseDTO.builder()
+                        .id(commentEntity.getId())
                         .userPK(commentEntity.getUserPK())
                         .profile(commentUserEntity.getProfile())
                         .content(commentEntity.getContent())
                         .createTime(commentEntity.getCreatedDate())
-//                        .updateTime(commentEntity.getLastModifiedDate())
                         .build();
 
                 comments.add(commentDTO);
@@ -277,7 +281,7 @@ public class ArticleServiceImpl implements ArticleService{
         // image 조회
         List<String> imgPaths = new ArrayList<>();
         List<ArticleImageEntity> articleImageEntityList = articleImageRepository.findAllByArticleId(articleId);
-        if(articleImageEntityList != null){
+        if(articleImageEntityList.size() != 0){
             for(ArticleImageEntity entity : articleImageEntityList){
                 imgPaths.add(entity.getImg());
             }
@@ -363,14 +367,19 @@ public class ArticleServiceImpl implements ArticleService{
 
         // 태크-게시물 테이블에 기존 태그 delete
         List<ArticleTagEntity> articleTagEntityList = articleTagRepository.findByArticleId(articleDTO.getArticleId());
-        for(ArticleTagEntity articleTagEntity : articleTagEntityList){
-            articleTagRepository.delete(articleTagEntity);
+        if(articleTagEntityList.size() != 0){
+            for(ArticleTagEntity articleTagEntity : articleTagEntityList){
+                articleTagRepository.delete(articleTagEntity);
+            }
         }
+
 
         // 이미지 테이블의 기존 정보 delete
         List<ArticleImageEntity> articleImageEntityList = articleImageRepository.findAllByArticleId(articleDTO.getArticleId());
-        for(ArticleImageEntity articleImageEntity : articleImageEntityList){
-            articleImageRepository.delete(articleImageEntity);
+        if(articleImageEntityList.size() != 0 ){
+            for(ArticleImageEntity articleImageEntity : articleImageEntityList){
+                articleImageRepository.delete(articleImageEntity);
+            }
         }
 
         // article 테이블에 update
@@ -495,19 +504,21 @@ public class ArticleServiceImpl implements ArticleService{
         List<ArticleEntity> articleEntityList = articleRepository.findUserSharedArticle(userPK, articleId);
 
         List<UserSharedArticleDTO> result = new ArrayList<>();
-        for(ArticleEntity entity : articleEntityList){
-            // 이미지 조회
-            List<ArticleImageEntity> aImageEntityList = articleImageRepository.findAllByArticleId(entity.getId());
+        if(articleEntityList.size() != 0){
+            for(ArticleEntity entity : articleEntityList){
+                // 이미지 조회
+                List<ArticleImageEntity> aImageEntityList = articleImageRepository.findAllByArticleId(entity.getId());
 
-            UserSharedArticleDTO dto = UserSharedArticleDTO.builder()
-                    .articleId(entity.getId())
-                    .userPK(entity.getUserPK())
-                    .nickName(userRepository.findById(userPK).orElseThrow().getNickName())
-                    .title(entity.getTitle())
-                    .img((aImageEntityList == null || aImageEntityList.size() == 0) ? null : aImageEntityList.get(0).getImg())
-                    .build();
+                UserSharedArticleDTO dto = UserSharedArticleDTO.builder()
+                        .articleId(entity.getId())
+                        .userPK(entity.getUserPK())
+                        .nickName(userRepository.findById(userPK).orElseThrow().getNickName())
+                        .title(entity.getTitle())
+                        .img((aImageEntityList == null || aImageEntityList.size() == 0) ? null : aImageEntityList.get(0).getImg())
+                        .build();
 
-            result.add(dto);
+                result.add(dto);
+            }
         }
 
         return result;
@@ -613,6 +624,4 @@ public class ArticleServiceImpl implements ArticleService{
         }
         return imgPath;
     }
-
-
 }
